crud TestDao

+++ src/main/java/by/a1qa/klimov/jdbc/dao/interfaces/ICrud.java
@@ -3,6 +3,6 @@ package by.a1qa.klimov.jdbc.dao.interfaces;
 public interface ICrud<T> {
     long create(T object);
     T read(long id);
-    long update(T object);
-    long delete(long id);
+    int update(T object);
+    int delete(long id);
 }
